rename some folders
